+++ lib/appsignal/event_formatter/faraday/request_formatter.rb
@@ -1,0 +1,18 @@
+module Appsignal
+  class EventFormatter
+    module Faraday
+      class RequestFormatter < Appsignal::EventFormatter
+        register 'request.faraday'
+
+        def format(payload)
+          http_method = payload[:method].to_s.upcase
+          uri = payload[:url]
+          [
+            "#{http_method} #{uri.scheme}://#{uri.host}",
+            "#{http_method} #{uri.scheme}://#{uri.host}#{uri.path}"
+          ]
+        end
+      end
+    end
+  end
+end