+++ lib/appsignal/hooks.rb
@@ -61,7 +61,11 @@ module Appsignal
 
       def extract_value(object_or_hash, field, default_value=nil, convert_to_s=false)
         value = if object_or_hash.respond_to?(:[])
-          object_or_hash[field]
+          begin
+            object_or_hash[field]
+          rescue NameError
+            nil
+          end
         elsif object_or_hash.respond_to?(field)
           object_or_hash.send(field)
         end || default_value